+++ exporter/awss3exporter/data_writer.go
@@ -17,6 +17,5 @@ package awss3exporter // import "github.com/open-telemetry/opentelemetry-collect
 import "context"
 
 type DataWriter interface {
-	WriteJson(buf []byte, config *Config) error
-	WriteParquet(metrics []*ParquetMetric, ctx context.Context, config *Config) error
+	WriteBuffer(ctx context.Context, buf []byte, config *Config, metadata string, format string) error
 }