+++ CHANGELOG.md
@@ -1861,6 +1861,7 @@ This version has been skipped.
 - `schemaprocessor`: Starting the initial work to allow from translating from semantic convention to another (#8371)
 - `saphanareceiver`: Added implementation of SAP HANA Metric Receiver (#8827)
 - `logstransformprocessor`: Add implementation of Logs Transform Processor (#9335)
+- `awss3exporter`: Add aws s3 exporter (#2835)
 
 ### 💡 Enhancements 💡